Clase BasiCalculator modificada

+++ src/main/java/co/com/sofka/app/calculator/BasicCalculator.java
@@ -22,10 +22,10 @@ public class BasicCalculator {
         return number1 * number2;
     }
 
-    public Double divide(Double number1, Double number2) {
+    public Double divide(Long number1, Long number2) {
         logger.info( "Dividing {} / {}", number1, number2 );
         try{
-            return number1 / number2;
+            return (double) (number1 / number2);
         } catch(Exception ex){
             ex.printStackTrace();
             throw new IllegalArgumentException("No se puede dividir entre cero.");
